intial refactoring
