+++ src/fileHandling/outputWriter/XYZWriter.cpp
@@ -5,7 +5,7 @@
  *      Author: eckhardw
  */
 
-#include "outputWriter/XYZWriter.h"
+#include "XYZWriter.h"
 #include <iomanip>
 #include <sstream>
