+++ src/fileHandling/outputWriter/vtk-unstructured.h
@@ -56,52 +56,52 @@
 //
 // End prologue.
 
-#include <xsd/cxx/config.hxx>
+#include "xsd/cxx/config.hxx"
 
 #if (XSD_INT_VERSION != 4000000L)
 #error XSD runtime version mismatch
 #endif
 
-#include <xsd/cxx/pre.hxx>
+#include "xsd/cxx/pre.hxx"
 
-#include <xsd/cxx/xml/char-utf8.hxx>
+#include "xsd/cxx/xml/char-utf8.hxx"
 
-#include <xsd/cxx/tree/exceptions.hxx>
-#include <xsd/cxx/tree/elements.hxx>
-#include <xsd/cxx/tree/types.hxx>
+#include "xsd/cxx/tree/exceptions.hxx"
+#include "xsd/cxx/tree/elements.hxx"
+#include "xsd/cxx/tree/types.hxx"
 
-#include <xsd/cxx/xml/error-handler.hxx>
+#include "xsd/cxx/xml/error-handler.hxx"
 
-#include <xsd/cxx/xml/dom/auto-ptr.hxx>
+#include "xsd/cxx/xml/dom/auto-ptr.hxx"
 
-#include <xsd/cxx/tree/parsing.hxx>
-#include <xsd/cxx/tree/parsing/byte.hxx>
-#include <xsd/cxx/tree/parsing/unsigned-byte.hxx>
-#include <xsd/cxx/tree/parsing/short.hxx>
-#include <xsd/cxx/tree/parsing/unsigned-short.hxx>
-#include <xsd/cxx/tree/parsing/int.hxx>
-#include <xsd/cxx/tree/parsing/unsigned-int.hxx>
-#include <xsd/cxx/tree/parsing/long.hxx>
-#include <xsd/cxx/tree/parsing/unsigned-long.hxx>
-#include <xsd/cxx/tree/parsing/boolean.hxx>
-#include <xsd/cxx/tree/parsing/float.hxx>
-#include <xsd/cxx/tree/parsing/double.hxx>
-#include <xsd/cxx/tree/parsing/decimal.hxx>
+#include "xsd/cxx/tree/parsing.hxx"
+#include "xsd/cxx/tree/parsing/byte.hxx"
+#include "xsd/cxx/tree/parsing/unsigned-byte.hxx"
+#include "xsd/cxx/tree/parsing/short.hxx"
+#include "xsd/cxx/tree/parsing/unsigned-short.hxx"
+#include "xsd/cxx/tree/parsing/int.hxx"
+#include "xsd/cxx/tree/parsing/unsigned-int.hxx"
+#include "xsd/cxx/tree/parsing/long.hxx"
+#include "xsd/cxx/tree/parsing/unsigned-long.hxx"
+#include "xsd/cxx/tree/parsing/boolean.hxx"
+#include "xsd/cxx/tree/parsing/float.hxx"
+#include "xsd/cxx/tree/parsing/double.hxx"
+#include "xsd/cxx/tree/parsing/decimal.hxx"
 
-#include <xsd/cxx/xml/dom/serialization-header.hxx>
-#include <xsd/cxx/tree/serialization.hxx>
-#include <xsd/cxx/tree/serialization/byte.hxx>
-#include <xsd/cxx/tree/serialization/unsigned-byte.hxx>
-#include <xsd/cxx/tree/serialization/short.hxx>
-#include <xsd/cxx/tree/serialization/unsigned-short.hxx>
-#include <xsd/cxx/tree/serialization/int.hxx>
-#include <xsd/cxx/tree/serialization/unsigned-int.hxx>
-#include <xsd/cxx/tree/serialization/long.hxx>
-#include <xsd/cxx/tree/serialization/unsigned-long.hxx>
-#include <xsd/cxx/tree/serialization/boolean.hxx>
-#include <xsd/cxx/tree/serialization/float.hxx>
-#include <xsd/cxx/tree/serialization/double.hxx>
-#include <xsd/cxx/tree/serialization/decimal.hxx>
+#include "xsd/cxx/xml/dom/serialization-header.hxx"
+#include "xsd/cxx/tree/serialization.hxx"
+#include "xsd/cxx/tree/serialization/byte.hxx"
+#include "xsd/cxx/tree/serialization/unsigned-byte.hxx"
+#include "xsd/cxx/tree/serialization/short.hxx"
+#include "xsd/cxx/tree/serialization/unsigned-short.hxx"
+#include "xsd/cxx/tree/serialization/int.hxx"
+#include "xsd/cxx/tree/serialization/unsigned-int.hxx"
+#include "xsd/cxx/tree/serialization/long.hxx"
+#include "xsd/cxx/tree/serialization/unsigned-long.hxx"
+#include "xsd/cxx/tree/serialization/boolean.hxx"
+#include "xsd/cxx/tree/serialization/float.hxx"
+#include "xsd/cxx/tree/serialization/double.hxx"
+#include "xsd/cxx/tree/serialization/decimal.hxx"
 
 /**
  * @brief C++ namespace for the %http://www.w3.org/2001/XMLSchema
@@ -633,14 +633,14 @@ class Cells;
 #include <algorithm> // std::binary_search
 #include <utility>   // std::move
 
-#include <xsd/cxx/xml/char-utf8.hxx>
+#include "xsd/cxx/xml/char-utf8.hxx"
 
-#include <xsd/cxx/tree/exceptions.hxx>
-#include <xsd/cxx/tree/elements.hxx>
-#include <xsd/cxx/tree/containers.hxx>
-#include <xsd/cxx/tree/list.hxx>
+#include "xsd/cxx/tree/exceptions.hxx"
+#include "xsd/cxx/tree/elements.hxx"
+#include "xsd/cxx/tree/containers.hxx"
+#include "xsd/cxx/tree/list.hxx"
 
-#include <xsd/cxx/xml/dom/parsing-header.hxx>
+#include "xsd/cxx/xml/dom/parsing-header.hxx"
 
 /**
  * @brief List class corresponding to the %DataArrayList_t
@@ -3332,7 +3332,7 @@ VTKFile (::xml_schema::dom::unique_ptr< ::xercesc::DOMDocument > d,
 #include <xercesc/dom/DOMErrorHandler.hpp>
 #include <xercesc/framework/XMLFormatter.hpp>
 
-#include <xsd/cxx/xml/dom/auto-ptr.hxx>
+#include "xsd/cxx/xml/dom/auto-ptr.hxx"
 
 void
 operator<< (::xercesc::DOMElement&, const DataArrayList_t&);
@@ -3543,7 +3543,7 @@ operator<< (::xercesc::DOMElement&, const Points&);
 void
 operator<< (::xercesc::DOMElement&, const Cells&);
 
-#include <xsd/cxx/post.hxx>
+#include "xsd/cxx/post.hxx"
 
 // Begin epilogue.
 //
